pyproject.toml
[project]
name = "necker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "tomli-w>=1.1.0",
    "openai>=1.0.0",
    "pyyaml>=6.0.0",
]

[dependency-groups]
dev = [
    "python-lsp-server>=1.13.1",
    "ruff>=0.14.2",
]
